scripts/public/requirements.txt
# Paper Experiments Requirements

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Deep learning
torch>=1.10.0
torchvision>=0.11.0

# CLIP models
open_clip_torch>=2.20.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.62.0
